Linted and splitted cci indicator into several files.
