+++ ts/Stock/Indicators/CCI/CCIOptions.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type {
+    SMAOptions,
+    SMAParamsOptions
+} from '../SMA/SMAOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface CCIOptions extends SMAOptions {
+    params?: CCIParamsOptions;
+}
+
+export interface CCIParamsOptions extends SMAParamsOptions {
+    // for inheritance
+}
+
+export default CCIOptions;
