+++ ts/Stock/Indicators/CCI/CCIPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type CCIIndicator from './CCIIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class CCIPoint extends SMAPoint {
+    public series: CCIIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default CCIPoint;
